do not add empty text tags
