+++ src/Potaka/BbCode/Tokenizer/Tokenizer.php
@@ -51,9 +51,11 @@ class Tokenizer
                 if ($tagText[0] === '/') {
                     $tagName = mb_strcut($tagText, 1);
                     if ($currentTag->getType() === $tagName) {
-                        $tmpTag = new Tag('text');
-                        $tmpTag->setText($bufferText);
-                        $currentTag->addTag($tmpTag);
+                        if ($bufferText !== '') {
+                            $tmpTag = new Tag('text');
+                            $tmpTag->setText($bufferText);
+                            $currentTag->addTag($tmpTag);
+                        }
                         $currentTag = $currentTag->getParent();
                     } else {
                         // ? add to bufferText if fail ?
@@ -78,66 +80,5 @@ class Tokenizer
         }
 
         return $this->rootTag;
-    }
-}
-
-class Tag
-{
-    private $tags = [];
-    private $type;
-    private $parent = null;
-    private $text;
-
-    public function __construct($type)
-    {
-        $this->type = $type;
-    }
-
-    public function getParent()
-    {
-        return $this->parent;
-    }
-
-    public function setParent(Tag $parent)
-    {
-        $this->parent = $parent;
-    }
-
-    public function addTag(Tag $tag)
-    {
-        $tag->setParent($this);
-        $this->tags[] = $tag;
-
-        return $this;
-    }
-
-    /**
-     *
-     * @param bool $reverse
-     * @return Tag[]
-     */
-    public function getTags($reverse = false)
-    {
-        if ($reverse) {
-            return array_reverse($this->tags);
-        }
-
-        return $this->tags;
-    }
-
-    public function getType()
-    {
-        return $this->type;
-    }
-
-    public function getText()
-    {
-        return $this->text;
-    }
-
-    public function setText($text)
-    {
-        $this->text = $text;
-        return $this;
     }
 }
